Add examples to use new runner of calibration pipeline
